requirements.txt
annotated-types==0.7.0
kafka-python==2.0.2
pydantic==2.9.2
pydantic_core==2.23.4
typing_extensions==4.12.2
